# docs/swagger.yaml
definitions:
...
  server.AuthorizationRequest:
    properties:
      action:
        description: |-
          Action being performed
          required: true
          example: read
        type: string
      context:
        additionalProperties: true
        description: |-
          Additional context for the authorization decision
          example: {"ip": "192.168.1.1", "time": "2024-01-01T12:00:00Z", "department": "engineering"}
        type: object
      resource:
        description: |-
          Resource being accessed
          required: true
          example: document456
        type: string
      subject:
        description: |-
          Subject performing the action (user, service, etc.)
          required: true
          example: user123
        type: string
    required:
    - action
    - resource
    - subject
    type: object
  server.AuthorizationResponse:
    properties:
      allowed:
        description: |-
          Whether the action is allowed
          example: true
        type: boolean
      decision_time_ms:
        description: |-
          Decision time in milliseconds
          example: 15
        type: integer
      policy_id:
        description: |-
          Policy that made the decision (optional)
          example: rbac-admin-policy
        type: string
      reason:
        description: |-
          Reason for the decision (optional)
          example: User has admin role with read permission
        type: string
    type: object
...
  server.CreatePolicyRequest:
    type: object
  server.CreatePolicyResponse:
    type: object
  server.CreatePolicyVersionRequest:
    type: object
  server.CreatePolicyVersionResponse:
    type: object
...
  server.ErrorResponse:
    properties:
      error:
        description: |-
          Error code
          example: validation_error
        type: string
      message:
        description: |-
          Error message
          example: Role name is required
        type: string
    type: object
...
  server.PolicyDetails:
    type: object
  server.PolicyVersionDetails:
    type: object
...
  server.UpdatePolicyRequest:
    type: object
  server.UpdatePolicyResponse:
    type: object
...
  server.ValidatePolicyRequest:
    type: object
  server.ValidatePolicyResponse:
    type: object
info:
  contact: {}
paths:
...
  /api/v1/authorize:
    post:
      consumes:
      - application/json
      description: Check if a subject is authorized to perform an action on a resource
      parameters:
      - description: Authorization request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.AuthorizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authorization decision
          schema:
            $ref: '#/definitions/server.AuthorizationResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Authorize an action
      tags:
      - 0-core-authorization
...
  /api/v1/policies:
    get:
      description: Get a list of all policies in the system
      parameters:
      - description: Filter by policy type (rbac, abac, rebac)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of policies
          schema:
            items:
              $ref: '#/definitions/server.PolicyDetails'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List all policies
      tags:
      - 1-core-policies
    post:
      consumes:
      - application/json
      description: Create a new policy that can support multiple access control models
      parameters:
      - description: Policy creation request
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/server.CreatePolicyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Policy created successfully
          schema:
            $ref: '#/definitions/server.CreatePolicyResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new policy
      tags:
      - 1-core-policies
  /api/v1/policies/{id}:
    delete:
      description: Delete an existing policy from the system
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Policy deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Delete a policy
      tags:
      - 1-core-policies
    put:
      consumes:
      - application/json
      description: Update an existing policy in the system
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: Policy update request
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/server.UpdatePolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Policy updated successfully
          schema:
            $ref: '#/definitions/server.UpdatePolicyResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Update a policy
      tags:
      - 1-core-policies
  /api/v1/policies/{id}/versions:
    get:
      description: Get a list of all versions for a specific policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of policy versions
          schema:
            items:
              $ref: '#/definitions/server.PolicyVersionDetails'
            type: array
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List policy versions
      tags:
      - policy-versions
    post:
      consumes:
      - application/json
      description: Create a new version of an existing policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: Policy version creation request
        in: body
        name: version
        required: true
        schema:
          $ref: '#/definitions/server.CreatePolicyVersionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Policy version created successfully
          schema:
            $ref: '#/definitions/server.CreatePolicyVersionResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new policy version
      tags:
      - policy-versions
  /api/v1/policies/validate:
    post:
      consumes:
      - application/json
      description: Validate the syntax and semantics of a policy
      parameters:
      - description: Policy validation request
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/server.ValidatePolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Policy validation result
          schema:
            $ref: '#/definitions/server.ValidatePolicyResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Validate a policy
      tags:
      - 1-core-policies
...
swagger: "2.0"